+++ src/flickr_photos_api/__init__.py
@@ -26,7 +26,7 @@ from .types import (
 )
 
 
-__version__ = "1.7.0"
+__version__ = "1.8.0"
 
 
 __all__ = [