+++ src/flickr_photos_api/api.py
@@ -25,10 +25,8 @@ from .exceptions import (
 )
 from .types import (
     CollectionOfPhotos,
-    DateTaken,
     GroupInfo,
     License,
-    LocationInfo,
     PhotosFromUrl,
     PhotosInAlbum,
     PhotosInGallery,
@@ -40,7 +38,7 @@ from .types import (
 from .utils import (
     parse_date_posted,
     parse_date_taken,
-    parse_date_taken_granularity,
+    parse_location,
     parse_safety_level,
     parse_sizes,
 )
@@ -306,35 +304,6 @@ class FlickrPhotosApi(BaseApi):
 
         return self.lookup_user_by_id(user_id=user_id)
 
-    def _get_date_taken(
-        self, *, value: str, granularity: str, unknown: bool
-    ) -> DateTaken | None:
-        # Note: we intentionally omit sending any 'date taken' information
-        # to callers if it's unknown.
-        #
-        # There will be a value in the API response, but if the taken date
-        # is unknown, it's defaulted to the date the photo was posted.
-        # See https://www.flickr.com/services/api/misc.dates.html
-        #
-        # This value isn't helpful to callers, so we omit it.  This reduces
-        # the risk of somebody skipping the ``unknown`` parameter and using
-        # the value in the wrong place.
-        if unknown:
-            return None
-
-        # This is a weird value I've seen returned on some videos; I'm
-        # not sure what it means, but it's not something we can interpret
-        # as a valid date, so we treat "date taken" as unknown even if
-        # the API thinks it knows it.
-        elif value.startswith("0000-"):
-            return None
-
-        else:
-            return {
-                "value": parse_date_taken(value),
-                "granularity": parse_date_taken_granularity(granularity),
-            }
-
     def get_single_photo(self, *, photo_id: str) -> SinglePhoto:
         """
         Look up the information for a single photo.
@@ -396,7 +365,7 @@ class FlickrPhotosApi(BaseApi):
 
         date_posted = parse_date_posted(dates["posted"])
 
-        date_taken = self._get_date_taken(
+        date_taken = parse_date_taken(
             value=dates["taken"],
             granularity=dates["takengranularity"],
             unknown=dates["takenunknown"] == "1",
@@ -477,30 +446,15 @@ class FlickrPhotosApi(BaseApi):
             assert t.text is not None
             tags.append(t.text)
 
+        # Get location information about the photo.
+        #
         # The <location> tag is only present in photos which have
         # location data; if the user hasn't made location available to
         # public users, it'll be missing.
-        #
-        # The accuracy parameter in the Flickr API response tells us
-        # the precision of the location information (15 November 2023):
-        #
-        #     Recorded accuracy level of the location information.
-        #     World level is 1, Country is ~3, Region ~6, City ~11, Street ~16.
-        #     Current range is 1-16.
-        #
-        # But some photos have an accuracy of 0!  It's unclear what this
-        # means or how we should map this -- lower numbers mean less accurate,
-        # so this location information might be completely useless.
-        #
-        # Discard it rather than risk writing bad data into Wikimedia.
         location_elem = photo_elem.find(path="location")
 
-        if location_elem is not None and location_elem.attrib["accuracy"] != "0":
-            location: LocationInfo | None = {
-                "latitude": float(location_elem.attrib["latitude"]),
-                "longitude": float(location_elem.attrib["longitude"]),
-                "accuracy": int(location_elem.attrib["accuracy"]),
-            }
+        if location_elem is not None:
+            location = parse_location(location_elem)
         else:
             location = None
 
@@ -596,11 +550,7 @@ class FlickrPhotosApi(BaseApi):
             # <photo> element to determine if there's actually location
             # information here, or if we're getting the defaults.
             if photo_elem.attrib.get("geo_is_public") == "1":
-                location: LocationInfo | None = {
-                    "latitude": float(photo_elem.attrib["latitude"]),
-                    "longitude": float(photo_elem.attrib["longitude"]),
-                    "accuracy": int(photo_elem.attrib["accuracy"]),
-                }
+                location = parse_location(photo_elem)
             else:
                 location = None
 
@@ -610,7 +560,7 @@ class FlickrPhotosApi(BaseApi):
                     "title": title,
                     "description": description,
                     "date_posted": parse_date_posted(photo_elem.attrib["dateupload"]),
-                    "date_taken": self._get_date_taken(
+                    "date_taken": parse_date_taken(
                         value=photo_elem.attrib["datetaken"],
                         granularity=photo_elem.attrib["datetakengranularity"],
                         unknown=photo_elem.attrib["datetakenunknown"] == "1",